:art: Split with implementations
